requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
click==7.1.2
cymem==2.0.5
easydict==1.9
en-core-web-sm==3.1.0
es-core-news-sm==3.1.0
xx-ent-wiki-sm==3.1.0
Flask==2.0.1
fold-to-ascii==1.0.2.post1
idna==3.2
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
nltk==3.6.2
numpy==1.21.0
pandas==1.3.1
pdfminer.six==20201018
pip==21.2.1
psutil==5.8.0
pyPdf==1.13
PyPDF2==1.26.0
python-dateutil==2.8.2
python-docx==0.8.11
regex==2021.7.6
requests==2.26.0
setuptools==47.1.0
six==1.16.0
spacy==3.1.0
textblob==0.15.3
tqdm==4.61.2
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1